Moved MSELoss definition to a header file
